ref: fold duplicated bodies in deprecator, shadowed-binding rule and indenters
